+++ CHANGELOG.md
@@ -108,3 +108,4 @@
   - Keycloak RoleService provider implementation
   - Keycloak GroupService provider implementation
   - Keycloak UserService provider implementation
+  - Keycloak Controller for User, Roles and Groups creation and list